MagicMetaModelService
