+++ src/main/kotlin/org/jetbrains/plugins/bsp/listeners/MyProjectManagerListener.kt
@@ -3,11 +3,11 @@ package org.jetbrains.plugins.bsp.listeners
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManagerListener
-import org.jetbrains.plugins.bsp.services.MyProjectService
+import org.jetbrains.plugins.bsp.services.MagicMetaModelService
 
 internal class MyProjectManagerListener : ProjectManagerListener {
 
   override fun projectOpened(project: Project) {
-    project.service<MyProjectService>()
+    project.service<MagicMetaModelService>()
   }
 }
