--- src/main/kotlin/org/jetbrains/plugins/bsp/services/MyApplicationService.kt
@@ -1,10 +1,0 @@
-package org.jetbrains.plugins.bsp.services
-
-import org.jetbrains.plugins.bsp.MyBundle
-
-class MyApplicationService {
-
-  init {
-    println(MyBundle.message("applicationService"))
-  }
-}
