+++ src/main/kotlin/org/jetbrains/plugins/bsp/startup/TestStartupActivity.kt
@@ -5,14 +5,8 @@ import com.intellij.openapi.application.runWriteAction
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ex.ProjectManagerEx
 import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.project.stateStore
 import com.intellij.projectImport.ProjectOpenProcessor
-import com.intellij.workspaceModel.ide.WorkspaceModel
-import com.intellij.workspaceModel.ide.getInstance
-import com.intellij.workspaceModel.storage.url.VirtualFileUrlManager
-import org.jetbrains.magicmetamodel.MagicMetaModel
-import org.jetbrains.magicmetamodel.MagicMetaModelProjectConfig
-import org.jetbrains.magicmetamodel.ProjectDetails
+import org.jetbrains.plugins.bsp.services.MagicMetaModelService
 import java.nio.file.Paths
 
 class TestStartupActivity : ProjectOpenProcessor() {
@@ -34,47 +28,8 @@ class TestStartupActivity : ProjectOpenProcessor() {
     val options = OpenProjectTask(isNewProject = true)
     val project = ProjectManagerEx.getInstanceEx().openProject(Paths.get(virtualFile.path), options)!!
 
-    val workspaceModel = WorkspaceModel.getInstance(project)
-
-    val virtualFileUrlManager = VirtualFileUrlManager.getInstance(project)
-    val projectBaseDir = project.stateStore.projectBasePath
-
-    val magicMetaModelProjectConfig = MagicMetaModelProjectConfig(workspaceModel, virtualFileUrlManager, projectBaseDir)
-
-    val projectDetails = ProjectDetails(
-      targetsId = SampleBSPProjectToImport.allTargetsIds,
-      targets = listOf(
-        SampleBSPProjectToImport.libATarget,
-        SampleBSPProjectToImport.appTarget,
-        SampleBSPProjectToImport.libBB2Target,
-        SampleBSPProjectToImport.libBBTarget,
-        SampleBSPProjectToImport.anotherAppTarget,
-        SampleBSPProjectToImport.libBTarget,
-      ),
-      sources = listOf(
-        SampleBSPProjectToImport.libBB2Sources,
-        SampleBSPProjectToImport.appSources,
-        SampleBSPProjectToImport.libASources,
-        SampleBSPProjectToImport.libBBSources,
-        SampleBSPProjectToImport.libBSources,
-        SampleBSPProjectToImport.anotherAppSources,
-      ),
-      resources = listOf(
-        SampleBSPProjectToImport.appResources,
-        SampleBSPProjectToImport.libBBResources,
-        SampleBSPProjectToImport.anotherAppResources,
-      ),
-      dependenciesSources = listOf(
-        SampleBSPProjectToImport.appDependenciesSources,
-        SampleBSPProjectToImport.libBBDependenciesSources,
-        SampleBSPProjectToImport.libBB2DependenciesSources,
-        SampleBSPProjectToImport.anotherAppDependenciesSources,
-        SampleBSPProjectToImport.libADependenciesSources,
-        SampleBSPProjectToImport.libBDependenciesSources,
-      ),
-    )
-
-    val magicMetaModel = MagicMetaModel.create(magicMetaModelProjectConfig, projectDetails)
+    val magicMetaModelService = MagicMetaModelService.getInstance(project)
+    val magicMetaModel = magicMetaModelService.magicMetaModel
 
     runWriteAction {
       magicMetaModel.loadDefaultTargets()
