+++ src/main/resources/META-INF/plugin.xml
@@ -8,6 +8,8 @@
     <depends>com.intellij.modules.platform</depends>
 
     <extensions defaultExtensionNs="com.intellij">
+      <projectService serviceImplementation="org.jetbrains.plugins.bsp.services.MagicMetaModelService"/>
+
       <projectOpenProcessor implementation="org.jetbrains.plugins.bsp.startup.TestStartupActivity"/>
